Create Ad-Clicks.py
A/ B testing is the statistical method to choose either option A or option B
Examples: 
1. if someone would click the green buttom or the blue button
2. Number of sales a website makes customers more likely to purchase based on age: 25 or 50

A/B testing where the outcome of interest (chosing A or B) is more categorical
An A/B testing is done using the chi square contingency hypotesis test
In order to run the test, we need to have the right sample size. A sample size calculator is based on three things
1. baseline conversion rate
2. minimum detectable effect aka minimum desirable lift
3. statistical significane threshold

BASELINE CONVERSION RATE
The A/B testing is used when there is the current version and another version we want to see if they are better. In order to compare the two versions, we have to use a metric. This is usually percentage.

MINIMUM DETECTABLE EFFECT AKA MINIMUM DESIREABLE LIFT
How much percentage would make us notice a difference, even if it is a small one?
In order to get the sample size, we need to know the minimum difference that would cause a change in the results
that minimum difference is the minimum detectable effect

HOW TO MAKE AN UNBIASED A/B TEST
1. do not rerun the test because we did not get the desired results. This will violate the ab testing hypothesis of chi squared contingency test unless the resulting pval is significant
2. do not stop the test till the sample size has been achieved, it can go to desired result to undesirable result when all the samples have been collected