Create LondonWeather.py
VARIENCE:
Consider two datasets:
dataset_one = [4, 2, 0, -2, -4]
dataset_two = [400, 200, 0, -200, -400]
From the two datasets shown, the mean, median and mode are all zero, if we were only reporting mean, median and mode, we would not be reporting any meaningful statistics.

To report more characteristics of a dataset, we can use varience. We can form one characteristics with the difference of a data and the mean.

Small Varience: When the difference is very small, that means that the values are closer to the mean
Large Varience: when the difference is very large, that means the values are very far from the mean

PROJECT:
You’re planning a trip to London and want to get a sense of the best time of the year to visit. Luckily, you got your hands on a dataset from 2015 that contains over 39,000 data points about weather conditions in London. Surely, with this much information, you can discover something useful about when to make your trip!