Create LifeExpectancyByCountry.py
Over the course of the past few centuries, technological and medical advancements have helped increase the life expectancy of humans. However, as of now, the average life expectancy of humans varies depending on what country you live in.

In this project, we will investigate a dataset containing information about the average life expectancy in 158 different countries. We will specifically look at how a country’s economic success might impact the life expectancy in that area.

QUARTILES:
An easy way to communicate what we see in a histogram is to split the data into four parts (all equally spaced)
By doing this, we can say if a new data point will lie in the first, second, third or fourth part of the histogram

QUANTILES:
When we want to evaluate the data, we can split the data into equal intervals. We can split them into 10 groups or we can split them into 100 groups. The group of data into various groups are called quantiles.
Quartile are a type of quantile where we split the data into 4 groups

INTERQUARTILE RANGE:
When we calculate the range of a dataset, they can be very large because of some outliers, as shown in the graph. Here, we can use IQR to identify the range of the majority of the data and ignores the tails of the histogram. The quantile range of 25% and 75% are used as the common interquartile range