Create Familiar.py
Welcome to Familiar, a startup in the new market of blood transfusion! You’ve joined the team because you appreciate the flexible hours and extremely intelligent team, but the overeager doorman welcoming you into the office is a nice way to start your workday (well, work-evening).

Familiar has fallen into some tough times lately, so you’re hoping to help them make some insights about their product and help move the needle (so to speak).

ONE SAMPLE T-TEST:
It compares a smaple mean to a hypothetical population mean. To do this, we make a null hypothesis, if the p-value is less than our threshold (0.05), we can reject the null hypothesis
How to avoid the Type I error: 
We can use a loop to calculate hte number of times the pval has been less than 0.05. Then calculate the percentage of unsuccessful pval and it should be around 5% to not reject the null hypothesis.

TWO SAMPLE T-TEST:
It compares two sets of numerical data. The null hypothesis is that the 2 observed sample comes from a population with the same mean.
The alternative hypothesis is thta the two observed sample from the population have different means. 
If the pval is less than 0.05, then we reject the null hypothesis and go with the alternative.
Dangers of Multiple T-Test:
1. More Chance of Type I error
2. Multiple T-Test will increase the chances of TypeI error since the probability increases
error = 1 - 0.95^n
where n is the number of times we run the test.

CHI SQAURED CONTINGENCY TEST:
They are useful to summerize two variables at the same time.
When calculating the expected contingency table, the observed and the expected table will look very different and we cannot make out if there is a relationship between the two  variables. We then calculate the Chi-Square of the observed and the expected.
Chi-Squared = sum((Observed - Expected)^2)
Output of the Chi-Squared Contingency Test:
The output is a single number for chi2
The interpretation of the chi-square number is dependant of the size of the table
For a table that is 2x2, a number greater than 4 would suggest an association of the numbers
We can also check on the threshold value (pval) from the test