Create FetchMaker.py
FICTIONAL COMPANY:
Congratulations! You’ve just started working at the hottest new tech startup, FetchMaker. FetchMaker’s mission is to match up prospective dog owners with their perfect pet. Data on thousands of adoptable dogs are in FetchMaker’s system, and it’s your job to analyze some of that data.

BINOM TEST:
Binomial tests are used for binary category data to compare a sample frequency to an expected population level probability

ANOVA:
When comparing more than two numberical datasets, one way to preserve a Type I error probability of 0.05 is to use ANOVA.
The ANOVA tests all the three sample with the same mean from the population and Null hypothesis.
The null hypothesis is that all the three samples have the same mean. Althernative hypothesis (if pval < 0.05) is when there is one pair with a different mean. We cannot make out which pair when using ANOVA.

TUKEY RANGE TEST:
When ANOVA runs an analysis with enough provision to avoid Type I error, although we cannot tell which pair shows the althernative hypothesis, this is where Tukey Range Test comes in.
The Tukey Range Test Runs all the tests simutaneously to preserve the type I error and will provide the conclusion of whether to reject the Null Hypothesis or not.

ASSUMPTIONS OF TTEST AND ANOVA:
1. The samples are normally distributed....ish. [bigger the sample, safer we are]
2. STD of the samples should be equal. [divide the 2 std to see if the ration is close to 1: ratio > 0.9 and ratio < 1.1]
3. The sampls must be independant. [understanding the data before running an analysis so that we can choose the right test]

CHI SQUARED CONTINGENCY TEST:
They are useful to summerize two variables at the same time.
When calculating the expected contingency table, the observed and the expected table will look very different and we cannot make out if there is a relationship between the two  variables. We then calculate the Chi-Square of the observed and the expected.
Chi-Squared = sum((Observed - Expected)^2)
Output of the Chi-Squared Contingency Test:
The output is a single number for chi2
The interpretation of the chi-square number is dependant of the size of the table
For a table that is 2x2, a number greater than 4 would suggest an association of the numbers
We can also check on the threshold value (pval) from the test