Create FarmBurg.py
FICITIONAL COMPANY
Brian is a Product Manager at FarmBurg, a company that makes a farming simulation social network game. In the FarmBurg game, you can plow, plant, and harvest different crops.
Today, you will be acting as Brian’s data analyst for an A/B Test that he has been conducting.

We want to see if the percent of Group A that purchased an upgrade package is significantly greater than p_clicks_099 (the percent of visitors who need to buy an upgrade package at $0.99 in order to make our target of $1,000).

CHI SQUARED CONTINGENCY TEST:
They are useful to summerize two variables at the same time.
When calculating the expected contingency table, the observed and the expected table will look very different and we cannot make out if there is a relationship between the two  variables. We then calculate the Chi-Square of the observed and the expected.
Chi-Squared = sum((Observed - Expected)^2)
Output of the Chi-Squared Contingency Test:
The output is a single number for chi2
The interpretation of the chi-square number is dependant of the size of the table
For a table that is 2x2, a number greater than 4 would suggest an association of the numbers
We can also check on the threshold value (pval) from the test

BINOM TEST:
Binomial tests are used for binary category data to compare a sample frequency to an expected population level probability.
Null hypothesis: true probability of purchae is the calculated number
Alternative Hypothesis: the probability is not the calculated number
If the pval is less than 0.05, then we reject the null hypothesis